requirements.txt
django==2.2
djangorestframework==3.9.2
nltk==3.4.5
vaderSentiment==3.2.1
textblob==0.15.3
